Define public API of `datalad_next.archive_operations`
See https://github.com/datalad/datalad-next/issues/613
